+++ src/app/app-routing.module.ts
@@ -1,0 +1,29 @@
+import { NgModule } from '@angular/core';
+import { RouterModule, Routes } from '@angular/router';
+
+import { ProductsListComponent } from './modules/products/components/product-list/products-list.component';
+import { ContactUsComponent } from './modules/contact-us/contact-us.component';
+
+const routes: Routes = [
+  {
+    path: 'products',
+    component: ProductsListComponent
+  },
+  {
+    path: 'contact-us',
+    component: ContactUsComponent
+  },
+  {
+    path: '',
+    redirectTo: '/products',
+    pathMatch: 'full'
+  }
+];
+
+@NgModule({
+  imports: [
+    RouterModule.forRoot(routes)
+  ],
+  exports: [RouterModule]
+})
+export class AppRoutingModule {}