+++ src/app/core/services/local-storage.service.ts
@@ -1,0 +1,22 @@
+import { Injectable } from '@angular/core';
+import { CoreModule } from '../core.module';
+
+@Injectable({
+  providedIn: CoreModule
+})
+export class LocalStorageService {
+
+  constructor() { }
+
+  setItem(name: string, value: string): void {
+    localStorage.setItem(name, value);
+  }
+
+  getItem(name: string): string {
+    return <string>localStorage.getItem(name);
+  }
+
+  removeItem(name: string): void {
+    localStorage.removeItem(name);
+  }
+}